refactor: extract parseRepoPath to src/utils/gitPathParser (#341)
Single canonical implementation. Replace 7 duplicate copies (one each
in GitService, RepoFileSyncService, RepoPullService, NoteGitHubSync,
CanvasGitHubSync, TodoGitHubSync, plus the re-exported one in
RepoFileBrowser) with imports from the shared utility.

All 4 consumers (NotesList, Explore, Home, MoveNoteDialog) now import
from utils/gitPathParser directly; the re-export from RepoFileBrowser
is removed.

Closes #341
