+++ src/utils/gitPathParser.ts
@@ -1,0 +1,17 @@
+// Canonical owner/repo parser. Accepts:
+//   facebook/react
+//   github.com/facebook/react
+//   https://github.com/facebook/react
+//   facebook/react.git
+export function parseRepoPath(repoPath: string): { owner: string; repo: string } | null {
+  const cleaned = repoPath
+    .replace(/^https?:\/\/github\.com\//, '')
+    .replace(/^github\.com\//, '')
+    .replace(/\.git$/, '')
+    .trim();
+  const parts = cleaned.split('/');
+  if (parts.length >= 2) {
+    return { owner: parts[0], repo: parts[1] };
+  }
+  return null;
+}
